move app to separate file
