+++ javascript/experiments/proxy-to-webpack-dev-server/client.js
@@ -1,26 +1,7 @@
-import { StrictMode, useState } from 'react';
+import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-const increment = 1;
-const decrement = -1;
-
-function App() {
-  const [count, setCount] = useState(0);
-
-  function changeBy(value) {
-    return () => setCount((previousCount) => previousCount + value);
-  }
-
-  return (
-    <div>
-      <p>Wohooo! My interactive piece of web page!</p>
-      <p>HMR works as well! Cool!</p>
-      <button onClick={changeBy(increment)}>Increment</button>
-      <button onClick={changeBy(decrement)}>Decrement</button>
-      <p>The current count is: {count}</p>
-    </div>
-  );
-}
+import { App } from './App';
 
 const root = document.querySelector('#my_root');
